Update to Enable SSH Key Connection to Raspberry Pi:
next push: optional ability to establish an ssh connection with the main device upon booting the pi: would likely be not recommended, as opening an ssh connection only when required for the execution of a specific task is more elegant and security conscious. Could be useful if you plan to do other things that rely on the open ssh link; maybe potential for API based pulling of real time data to be shown in the gui python program, if that's possible.